refactor(tools): split the fresh-init package-manager table
